% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_cache_reset.R
\name{lt_cache_reset}
\alias{lt_cache_reset}
\title{Reset the cached PostgREST schema}
\usage{
lt_cache_reset()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Clears any cached API schema data, forcing the next call to \code{\link[=lt_tables]{lt_tables()}} or
\code{\link[=lt_schema]{lt_schema()}} to re-fetch from the API.
}
\examples{
\dontrun{
  lt_cache_reset()
}
}
